+++ src/test/testData/annotator/AnnotatorFile3.glsl
@@ -1,0 +1,3 @@
+int a = <error descr="Incompatible types in initialization (and no available implicit conversion).">normalize(2)</error>;
+float a = normalize(2);
+float a = normalize(2, 3);