+++ src/main/kotlin/glsl/plugin/psi/named/variables/GlslNamedFunctionDeclarator.kt
@@ -6,25 +6,23 @@ import com.intellij.lang.ASTNode
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.psi.util.parentOfType
 import glsl.plugin.code.highlighting.GlslTextAttributes
-import glsl.plugin.psi.named.GlslNamedElement
 import glsl.plugin.psi.named.GlslNamedType
 import glsl.plugin.psi.named.GlslNamedVariableImpl
 import glsl.plugin.utils.GlslUtils
-import glsl.psi.interfaces.GlslFunctionHeader
-import glsl.psi.interfaces.GlslFunctionPrototype
+import glsl.psi.interfaces.GlslFunctionDeclarator
 import glsl.psi.interfaces.GlslVariableIdentifier
 import javax.swing.Icon
 
 /**
  *
  */
-abstract class GlslNamedFunctionHeader(node: ASTNode) : GlslNamedVariableImpl(node) {
+abstract class GlslNamedFunctionDeclarator(node: ASTNode) : GlslNamedVariableImpl(node) {
 
     /**
      *
      */
-    override fun getPsi(): GlslFunctionHeader {
-        return this as GlslFunctionHeader
+    override fun getPsi(): GlslFunctionDeclarator {
+        return this as GlslFunctionDeclarator
     }
 
     /**
@@ -59,15 +57,7 @@ abstract class GlslNamedFunctionHeader(node: ASTNode) : GlslNamedVariableImpl(no
      *
      */
     override fun getLookupElement(returnTypeText: String?): LookupElement? {
-        val functionPrototype = getPsi().parentOfType<GlslFunctionPrototype>() ?: return null
+        val functionPrototype = getPsi().parentOfType<GlslFunctionDeclarator>() ?: return null
         return GlslUtils.getFunctionLookupElement(functionPrototype, getLookupIcon())
-    }
-
-    /**
-     *
-     */
-    fun getParameterDeclarators(): List<GlslNamedElement> {
-        val funcPrototype = getPsi().parent as GlslFunctionPrototype? ?: return emptyList()
-        return funcPrototype.funcHeaderWithParams?.parameterDeclaratorList ?: emptyList()
     }
 }