+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # GLSL Plugin Changelog
 
+## [1.1.3]
+
 ## [1.1.2]
 ### Added
 - Autocompletion for include statement (not complete)